requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings[yaml]>=2.3.0

# HTTP
httpx[asyncio]>=0.25.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Retry
tenacity>=8.2.0

# Logging
loguru>=0.7.0

# TTS
edge-tts>=6.1.0

# Video/Audio processing
ffmpeg-python>=0.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
